Report correct signal for halted status
Previously we would just send dcsr.cause straight to GDB as a signal.
Being 3 when using haltreq, this was interpreted as SIGQUIT, so upon
continue we would get a C03 rather than a plain c, and we don't support
the former. SIGINT/SIGTERM are the two standard ones we should be using
(and are what OpenOCD uses for these cases) to ensure we only have to
deal with the basic c command.