Allow poll loops to be quickly preempted
The poll loops have a 1000000 poll timeout (falsely claimed to be usecs,
but it's rather greater than that in reality due to the time taken to
poll the debug module) before they return up and allow the file
descriptors to be checked for input again. In general, we don't allow
new GDB commands to preempt, only shutdown requests from the host
software, with the exception being for 'c' commands that must allow ^C
packets to halt the core in a timely manner.