Fixed a race condition in gdbstub_be (detail follows).
On GDB 'continue' we do:
- gdbstub_be_continue() which (A) DMI-writes DMCONTROL with RESUMEREQ
- gdbstub_be_get_stop_reason() which (B) polls, i.e., DMI-reads
    DMSTATUS for ALLHALTED, and then (C) DMI-reads DCSR for CAUSE.

Problem: if (B) is too quick after (A), CPU may not yet have responded
'started' to Debug Module; we falsely see ALLHALTED and move on to
(C). When we do (C) the CPU is running and the DCSR-read therefore
returns an error (not allowed to read unless stopped in Debug Mode).

Solution: In gdbstub_be_continue(), after (A), insert (A.1) poll,
i.e., DMI-read DMSTATUS for ALLRUNNING before returning, so the poll
(B) is guaranteed to be tried after Debug Module knows the CPU is
running.